requirements.txt
uvicorn
fastapi
python-dotenv
jinja2
python-multipart
pydantic
sqlalchemy
mysql-connector-python
python-jose
passlib
